# Switch Serde classes to act on STRUCT not GenericRow.