--- a/ksql-engine/src/main/java/io/confluent/ksql/util/PersistentQueryMetadata.java
+++ b/ksql-engine/src/main/java/io/confluent/ksql/util/PersistentQueryMetadata.java
@@ -104,7 +104,7 @@ public class PersistentQueryMetadata extends QueryMetadata {
   }
 
   public Format getResultTopicSerde() {
-    return resultTopic.getKsqlTopicSerDe().getSerDe();
+    return resultTopic.getKsqlTopicSerDe().getFormat();
   }
 
   public String getSchemasDescription() {