--- a/ksql-engine/src/test/java/io/confluent/ksql/EndToEndEngineTestUtil.java
+++ b/ksql-engine/src/test/java/io/confluent/ksql/EndToEndEngineTestUtil.java
@@ -1136,9 +1136,11 @@ final class EndToEndEngineTestUtil {
       case BOOLEAN:
         return spec;
       case ARRAY:
-        return ((List)spec).stream()
+        final List<?> list= ((List<?>) spec).stream()
             .map(o -> valueSpecToAvro(o, schema.getElementType()))
             .collect(Collectors.toList());
+
+        return new GenericData.Array(schema, list);
       case MAP:
         return ((Map<Object, Object>)spec).entrySet().stream().collect(
             Collectors.toMap(