--- a/ksql-rest-app/src/main/java/io/confluent/ksql/rest/entity/KsqlTopicInfo.java
+++ b/ksql-rest-app/src/main/java/io/confluent/ksql/rest/entity/KsqlTopicInfo.java
@@ -45,7 +45,7 @@ public class KsqlTopicInfo {
     this(
         ksqlTopic.getKsqlTopicName(),
         ksqlTopic.getKafkaTopicName(),
-        ksqlTopic.getKsqlTopicSerDe().getSerDe()
+        ksqlTopic.getKsqlTopicSerDe().getFormat()
     );
   }