--- a/ksql-rest-app/src/main/java/io/confluent/ksql/rest/util/ClusterTerminator.java
+++ b/ksql-rest-app/src/main/java/io/confluent/ksql/rest/util/ClusterTerminator.java
@@ -122,7 +122,7 @@ public class ClusterTerminator {
 
   private boolean hasAvroSource(final String topicName) {
     return ksqlEngine.getMetaStore().getSourcesForKafkaTopic(topicName).stream()
-        .anyMatch(dataSource -> dataSource.getKsqlTopicSerde().getSerDe() == Format.AVRO);
+        .anyMatch(dataSource -> dataSource.getKsqlTopicSerde().getFormat() == Format.AVRO);
   }
 
   private Stream<String> filterNonExistingSubjects(final Stream<String> subjects) {