… docs/changelog.rst
… ksql-benchmark/src/main/java/io/confluent/ksql/benchmark/SerdeBenchmark.java
… ksql-engine/src/test/java/io/confluent/ksql/QueryTranslationTest.java
… ksql-engine/src/test/java/io/confluent/ksql/analyzer/QueryAnalyzerTest.java
… ksql-engine/src/test/java/io/confluent/ksql/engine/KsqlEngineTest.java
… ksql-engine/src/test/java/io/confluent/ksql/integration/IntegrationTestHarness.java
… ksql-engine/src/test/java/io/confluent/ksql/physical/KsqlQueryBuilderTest.java
… ksql-engine/src/test/java/io/confluent/ksql/physical/PhysicalPlanBuilderTest.java
… ksql-engine/src/test/java/io/confluent/ksql/structured/SchemaKGroupedTableTest.java
… ksql-engine/src/test/java/io/confluent/ksql/structured/SchemaKStreamTest.java
… ksql-engine/src/test/java/io/confluent/ksql/structured/SchemaKTableTest.java
… ksql-engine/src/test/java/io/confluent/ksql/util/AvroUtilTest.java
… ksql-engine/src/test/java/io/confluent/ksql/util/TopicConsumer.java
… ksql-engine/src/test/java/io/confluent/ksql/util/TopicProducer.java
… ksql-engine/src/test/resources/query-validation-tests/README.md
… ksql-engine/src/test/resources/query-validation-tests/schemas.json
… ksql-functional-tests/README.md
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/model/FieldNode.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/model/KeyFieldNode.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/model/SourceNode.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/model/TestCaseNode.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/model/TopicNode.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/serde/avro/AvroSerdeSupplier.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/serde/avro/ValueSpecAvroSerdeSupplier.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/tools/Record.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/tools/TestCase.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/tools/Topic.java
… ksql-functional-tests/src/main/java/io/confluent/ksql/test/tools/exceptions/KsqlExpectedException.java
… ksql-functional-tests/src/test/java/io/confluent/ksql/test/tools/DataSourceMatchersTest.java
… ksql-functional-tests/src/test/java/io/confluent/ksql/test/tools/RecordTest.java