### Description 

The structured key work requires KSQL to be able to deserialize the key as types other than just `String`.  At the moment the KSQL serde classes for the different formats, (`JSON`, `AVRO` & `DELIMITED`), (de)serialize `GenericRow`.  The key will not be a generic row, but a `Struct`.

Rather than write a load of serde classes to handle the key `Struct` _alongside_ the existing serde classes for `GenericRow` this PR converts the existing serde classes to (de)serialize `Struct`, and then wraps this in `GenericRowSerde` which does the conversion between `Struct` and `GenericRow`.  This will allow future work to utilise these same serde classes for the key.

The initial phase of the structured key work will see KSQL support the full set of primitive keys, (`INT`, `BIGINT`, `DOUBLE`, etc). This means the likely use-cases will see keys that are just primitives, i.e. _not_ a struct with a single field, but just a primitive.

While repurposing the serde classes I've also extended them so that the deserializer classes support reading a serialized primitive type into a `Struct` if the schema only has one field, i.e. KSQL can now deserialize values, (and later keys), where there is no JSON document or Avro record, but just a primitive type, array or map. There are test cases showcasing this in `schemas.json`.

This fixes #1351 along the way.

### Testing done 

Extensive testing added.

Performance tests run:

Before changes:

```
Benchmark                   (schemaName)  (serializationFormat)  Mode  Cnt   Score   Error  Units
SerdeBenchmark.deserialize   impressions                   JSON  avgt    9   1.784 ± 0.333  us/op
SerdeBenchmark.deserialize   impressions                   Avro  avgt    9   2.676 ± 0.131  us/op
SerdeBenchmark.deserialize       metrics                   JSON  avgt    9   9.148 ± 0.679  us/op
SerdeBenchmark.deserialize       metrics                   Avro  avgt    9  10.045 ± 0.240  us/op
SerdeBenchmark.serialize     impressions                   JSON  avgt    9   0.746 ± 0.042  us/op
SerdeBenchmark.serialize     impressions                   Avro  avgt    9   1.658 ± 0.057  us/op
SerdeBenchmark.serialize         metrics                   JSON  avgt    9   3.274 ± 0.247  us/op
SerdeBenchmark.serialize         metrics                   Avro  avgt    9   6.201 ± 0.258  us/op
```

With changes:

```
Benchmark                   (schemaName)  (serializationFormat)  Mode  Cnt  Score   Error  Units
SerdeBenchmark.deserialize   impressions                   JSON  avgt    9  1.749 ± 0.135  us/op
SerdeBenchmark.deserialize   impressions                   Avro  avgt    9  2.519 ± 0.082  us/op
SerdeBenchmark.deserialize       metrics                   JSON  avgt    9  8.881 ± 0.186  us/op
SerdeBenchmark.deserialize       metrics                   Avro  avgt    9  9.691 ± 0.201  us/op
SerdeBenchmark.serialize     impressions                   JSON  avgt    9  0.696 ± 0.016  us/op
SerdeBenchmark.serialize     impressions                   Avro  avgt    9  1.709 ± 0.047  us/op
SerdeBenchmark.serialize         metrics                   JSON  avgt    9  3.203 ± 0.310  us/op
SerdeBenchmark.serialize         metrics                   Avro  avgt    9  6.367 ± 0.102  us/op
```

### Reviewer checklist
- [ ] Ensure docs are updated if necessary. (eg. if a user visible feature is being added or changed).
- [ ] Ensure relevant issues are linked (description should include text like "Fixes #<issue number>")